shopping_list_api/pyproject.toml
[tool.poetry]
name = "shopping-list-api"
version = "0.1.0"
description = ""
authors = ["Jai Jain <jai.jain@acsicorp.com>"]
readme = "README.md"
packages = [{include = "shopping_list_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-spectacular = "^0.26.3"
django-guardian = "^2.4.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
werkzeug = "^2.3.6"
tox = "3.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
